is deletion works?
